migrate staking sync to new arch
