+++ modules/actions/pool/staking/abi/ChildChainGaugeV1.json
