+++ modules/actions/pool/staking/abi/ChildChainGaugeV2.json
