+++ modules/actions/pool/staking/abi/EmissionCurve.json
