+++ modules/network/fantom.ts
@@ -4,10 +4,6 @@ import { BoostedPoolAprService } from '../pool/lib/apr-data-sources/nested-pool-
 import { SwapFeeAprService } from '../pool/lib/apr-data-sources/swap-fee-apr.service';
 import { MasterchefFarmAprService } from '../pool/lib/apr-data-sources/fantom/masterchef-farm-apr.service';
 import { ReliquaryFarmAprService } from '../pool/lib/apr-data-sources/fantom/reliquary-farm-apr.service';
-import { MasterChefStakingService } from '../pool/lib/staking/master-chef-staking.service';
-import { masterchefService } from '../subgraphs/masterchef-subgraph/masterchef.service';
-import { ReliquaryStakingService } from '../pool/lib/staking/reliquary-staking.service';
-import { reliquarySubgraphService } from '../subgraphs/reliquary-subgraph/reliquary.service';
 import { UserSyncMasterchefFarmBalanceService } from '../user/lib/user-sync-masterchef-farm-balance.service';
 import { UserSyncReliquaryFarmBalanceService } from '../user/lib/user-sync-reliquary-farm-balance.service';
 import { every } from '../../worker/intervals';
@@ -31,10 +27,6 @@ export const fantomNetworkConfig: NetworkConfig = {
         new MasterchefFarmAprService(fantomNetworkData.beets!.address),
         new ReliquaryFarmAprService(fantomNetworkData.beets!.address),
         new BeetswarsGaugeVotingAprService(),
-    ],
-    poolStakingServices: [
-        new MasterChefStakingService(masterchefService, fantomNetworkData.masterchef!.excludedFarmIds),
-        new ReliquaryStakingService(fantomNetworkData.reliquary!.address, reliquarySubgraphService),
     ],
     userStakedBalanceServices: [
         new UserSyncMasterchefFarmBalanceService(
