+++ modules/network/network-config-types.ts
@@ -1,8 +1,7 @@
 import type { Chain } from '@prisma/client';
 import type { BigNumber } from 'ethers';
-import type { PoolAprService, PoolStakingService } from '../pool/pool-types';
+import type { PoolAprService } from '../pool/pool-types';
 import type { UserStakedBalanceService } from '../user/user-types';
-import type { TokenPriceHandler } from '../token/token-types';
 import type { BaseProvider } from '@ethersproject/providers';
 import type { GqlChain } from '../../schema';
 import type { ContentService } from '../content/content-types';
@@ -13,7 +12,6 @@ import { SftmxSubgraphService } from '../sources/subgraphs/sftmx-subgraph/sftmx.
 export interface NetworkConfig {
     data: NetworkData;
     contentService: ContentService;
-    poolStakingServices: PoolStakingService[];
     poolAprServices: PoolAprService[];
     userStakedBalanceServices: UserStakedBalanceService[];
     provider: BaseProvider;
