--- modules/pool/lib/staking/index.ts
@@ -1,3 +1,0 @@
-export * from './gauge-staking.service';
-export * from './master-chef-staking.service';
-export * from './reliquary-staking.service';
