+++ modules/subgraphs/masterchef-subgraph/masterchef.service.ts
@@ -12,18 +12,16 @@ import {
     QueryMasterChefsArgs,
 } from './generated/masterchef-subgraph-types';
 import { subgraphLoadAll } from '../subgraph-util';
-import { twentyFourHoursInMs } from '../../common/time';
 import { Cache, CacheClass } from 'memory-cache';
 import { GraphQLClient } from 'graphql-request';
-import { networkContext } from '../../network/network-context.service';
-
-const ALL_FARM_USERS_CACHE_KEY = `masterchef-all-farm-users`;
 
 export class MasterchefSubgraphService {
     private readonly cache: CacheClass<string, any>;
+    private sdk: ReturnType<typeof getSdk>;
 
-    constructor() {
+    constructor(subgraphUrl: string) {
         this.cache = new Cache<string, any>();
+        this.sdk = getSdk(new GraphQLClient(subgraphUrl));
     }
 
     public async getMetadata() {
@@ -55,22 +53,6 @@ export class MasterchefSubgraphService {
         return this.sdk.MasterchefUsers(args);
     }
 
-    public async getFarmUsersAtBlock(address: string, block: number): Promise<FarmUserFragment[]> {
-        const cachedUsers = this.cache.get(`${ALL_FARM_USERS_CACHE_KEY}:${networkContext.chainId}:${block}`) as
-            | FarmUserFragment[]
-            | null;
-
-        if (cachedUsers) {
-            return cachedUsers.filter((user) => user.address === address) || null;
-        }
-
-        const users = await this.getAllFarmUsers({ block: { number: block } });
-
-        this.cache.put(`${ALL_FARM_USERS_CACHE_KEY}:${networkContext.chainId}:${block}`, users, twentyFourHoursInMs);
-
-        return users.filter((user) => user.id === address);
-    }
-
     public async getAllFarms(args: MasterchefFarmsQueryVariables): Promise<FarmFragment[]> {
         return subgraphLoadAll<FarmFragment>(this.sdk.MasterchefFarms, 'farms', args);
     }
@@ -86,12 +68,4 @@ export class MasterchefSubgraphService {
     public getFarmForPoolAddress(poolAddress: string, farms: FarmFragment[]): FarmFragment | null {
         return farms.find((farm) => farm.pair.toLowerCase() === poolAddress.toLowerCase()) || null;
     }
-
-    public get sdk() {
-        const client = new GraphQLClient(networkContext.data.subgraphs.masterchef!);
-
-        return getSdk(client);
-    }
 }
-
-export const masterchefService = new MasterchefSubgraphService();
